requirements.txt
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-filters==1.0.0.dev0
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
